Divided tests for navigation into two groups
